Remove unnecessary register service intereface
